refactor: test 공통 코드 리팩터링
